+++ y-2023/src/General/Day.java
@@ -2,6 +2,6 @@ package General;
 
 public interface Day {
     void loadData(String filename);
-    int part1();
-    int part2();
+    long part1();
+    long part2();
 }
